refactor(tracing): promote depydantic'd files from vendor/ to private modules
The two forked-and-tuned files no longer match the 'vendored copy kept
in sync with upstream' framing — they are ops_tracing's own modules now.
Move them next to the other private modules and rename to the project's
underscore convention:

  vendor/charms/.../certificate_transfer.py -> _cert_transfer_models.py
  vendor/charms/.../tracing.py              -> _tracing_models.py
  test/test_vendor_models.py                 -> test/test_models.py

vendor/otlp_json/ stays where it is (still genuinely vendored).
The fork-marker headers ('Forked from … de-pydantic'd … do not re-sync')
are unchanged — the file is still a fork, the path is just honest now.

Tests: 26 passed, 1 skipped (same as 993c942 baseline).
